+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [master](https://github.com/arangodb/go-driver/tree/master) (N/A)
 - Switch to Go 1.20.11
 - Switch to Go 1.21.5
+- Disable AF mode in tests (not supported since 3.12)
 
 ## [1.6.1](https://github.com/arangodb/go-driver/tree/v1.6.1) (2023-10-31)
 - Add support for getting license