+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [master](https://github.com/arangodb/go-driver/tree/master) (N/A)
+
+## [1.6.2](https://github.com/arangodb/go-driver/tree/v1.6.2) (2024-04-02)
 - Switch to Go 1.20.11
 - Switch to Go 1.21.5
 - Disable AF mode in tests (not supported since 3.12)